+++ CHANGELOG.md
@@ -6,7 +6,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
 
+Added:
 
+* New fanciness: Openshift WebHook
 
 
 ## 0.4.0 - 2016-11-10